added authorization helper and implemented for deleting builds

+++ app/api/builds/[id]/route.ts
@@ -1,3 +1,4 @@
+import { handleAuthToken } from "@/utils/handleAuthToken";
 import { sql } from "@vercel/postgres";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -28,13 +29,20 @@ export async function PUT(req: NextRequest, {params: {id}}: Props) {
 
 export async function DELETE(req: NextRequest, {params: {id}}: Props) {
   try {
-    if (!id) NextResponse.json({"error": "Missing id"}, {status: 400});
-    
-    await sql`DELETE FROM builds WHERE id=${id}`
+    const authToken = req.headers.get("Authorization");
+    console.log(authToken);
+    if (!id || !authToken) return NextResponse.json({"error": "Missing id or access token"}, {status: 400});
 
-    return NextResponse.json({"message": "Delete successful"}, {status: 200});
-}
-catch (error) {
+    const uid = await (await sql`SELECT uid FROM builds WHERE id=${id}`).rows[0].uid
+
+    if (authToken && await handleAuthToken(authToken, uid)) {
+      await sql`DELETE FROM builds WHERE id=${id}`
+      return NextResponse.json({"message": "Delete successful"}, {status: 200});
+    }
+    
+    return NextResponse.json({"error": "Access Token is invalid."}, {status: 500});
+  }
+  catch (error) {
     return NextResponse.json({error}, {status: 500});
-}
+  }
 }
